Added OptionsBase to reduce the code repetition
