+++ db-client-java/src/main/java/com/eventstore/dbclient/AppendToStreamOptions.java
@@ -1,47 +1,14 @@
 package com.eventstore.dbclient;
 
-import io.grpc.Metadata;
-
-public class AppendToStreamOptions {
-    private Timeouts timeouts;
-    private final ConnectionMetadata metadata;
-
+public class AppendToStreamOptions extends OptionsBase<AppendToStreamOptions> {
     private ExpectedRevision expectedRevision;
 
     private AppendToStreamOptions() {
-        this.timeouts = Timeouts.DEFAULT;
-        this.metadata = new ConnectionMetadata();
-
         this.expectedRevision = ExpectedRevision.ANY;
     }
 
     public static AppendToStreamOptions get() {
         return new AppendToStreamOptions();
-    }
-
-    public Timeouts getTimeouts() {
-        return this.timeouts;
-    }
-
-    public AppendToStreamOptions timeouts(Timeouts timeouts) {
-        this.timeouts = timeouts;
-        return this;
-    }
-
-    public Metadata getMetadata() {
-        return this.metadata.build();
-    }
-
-    public AppendToStreamOptions authenticated(UserCredentials credentials) {
-        if(credentials == null)
-            return this;
-
-        this.metadata.authenticated(credentials);
-        return this;
-    }
-
-    public boolean hasUserCredentials() {
-        return this.metadata.hasUserCredentials();
     }
 
     public ExpectedRevision getExpectedRevision() {
