+++ db-client-java/src/main/java/com/eventstore/dbclient/CreateContinuousProjectionOptions.java
@@ -1,34 +1,14 @@
 package com.eventstore.dbclient;
 
-import io.grpc.Metadata;
-
-public class CreateContinuousProjectionOptions {
-    private final ConnectionMetadata metadata;
+public class CreateContinuousProjectionOptions extends OptionsBase<CreateContinuousProjectionOptions> {
     private boolean trackEmittedStreams;
 
     private CreateContinuousProjectionOptions() {
-        this.metadata = new ConnectionMetadata();
         this.trackEmittedStreams = false;
     }
 
     public static CreateContinuousProjectionOptions get() {
         return new CreateContinuousProjectionOptions();
-    }
-
-    public Metadata getMetadata() {
-        return this.metadata.build();
-    }
-
-    public CreateContinuousProjectionOptions authenticated(UserCredentials credentials) {
-        if(credentials == null)
-            return this;
-
-        this.metadata.authenticated(credentials);
-        return this;
-    }
-
-    public boolean hasUserCredentials() {
-        return this.metadata.hasUserCredentials();
     }
 
     public boolean getTrackEmittedStreams() {
