+++ db-client-java/src/main/java/com/eventstore/dbclient/CreateOneTimeProjectionOptions.java
@@ -1,31 +1,7 @@
 package com.eventstore.dbclient;
 
-import io.grpc.Metadata;
-
-public class CreateOneTimeProjectionOptions {
-    private final ConnectionMetadata metadata;
-
-    private CreateOneTimeProjectionOptions() {
-        this.metadata = new ConnectionMetadata();
-    }
-
+public class CreateOneTimeProjectionOptions extends OptionsBase<CreateOneTimeProjectionOptions> {
     public static CreateOneTimeProjectionOptions get() {
         return new CreateOneTimeProjectionOptions();
-    }
-
-    public Metadata getMetadata() {
-        return this.metadata.build();
-    }
-
-    public CreateOneTimeProjectionOptions authenticated(UserCredentials credentials) {
-        if(credentials == null)
-            return this;
-
-        this.metadata.authenticated(credentials);
-        return this;
-    }
-
-    public boolean hasUserCredentials() {
-        return this.metadata.hasUserCredentials();
     }
 }
