+++ db-client-java/src/main/java/com/eventstore/dbclient/CreatePersistentSubscriptionOptions.java
@@ -1,34 +1,14 @@
 package com.eventstore.dbclient;
 
-import io.grpc.Metadata;
-
-public class CreatePersistentSubscriptionOptions {
-    private final ConnectionMetadata metadata;
+public class CreatePersistentSubscriptionOptions extends OptionsBase<CreatePersistentSubscriptionOptions> {
     private PersistentSubscriptionSettings settings;
 
     private CreatePersistentSubscriptionOptions() {
-        this.metadata = new ConnectionMetadata();
         this.settings = PersistentSubscriptionSettings.builder().build();
     }
 
     public static CreatePersistentSubscriptionOptions get() {
         return new CreatePersistentSubscriptionOptions();
-    }
-
-    public Metadata getMetadata() {
-        return this.metadata.build();
-    }
-
-    public CreatePersistentSubscriptionOptions authenticated(UserCredentials credentials) {
-        if(credentials == null)
-            return this;
-
-        this.metadata.authenticated(credentials);
-        return this;
-    }
-
-    public boolean hasUserCredentials() {
-        return this.metadata.hasUserCredentials();
     }
 
     public CreatePersistentSubscriptionOptions settings(PersistentSubscriptionSettings settings) {
