+++ db-client-java/src/main/java/com/eventstore/dbclient/DeletePersistentSubscriptionOptions.java
@@ -1,31 +1,7 @@
 package com.eventstore.dbclient;
 
-import io.grpc.Metadata;
-
-public class DeletePersistentSubscriptionOptions {
-    private final ConnectionMetadata metadata;
-
-    private DeletePersistentSubscriptionOptions() {
-        this.metadata = new ConnectionMetadata();
-    }
-
+public class DeletePersistentSubscriptionOptions extends OptionsBase<DeletePersistentSubscriptionOptions> {
     public static DeletePersistentSubscriptionOptions get() {
         return new DeletePersistentSubscriptionOptions();
-    }
-
-    public Metadata getMetadata() {
-        return this.metadata.build();
-    }
-
-    public DeletePersistentSubscriptionOptions authenticated(UserCredentials credentials) {
-        if(credentials == null)
-            return this;
-
-        this.metadata.authenticated(credentials);
-        return this;
-    }
-
-    public boolean hasUserCredentials() {
-        return this.metadata.hasUserCredentials();
     }
 }
