+++ db-client-java/src/main/java/com/eventstore/dbclient/UpdatePersistentSubscriptionOptions.java
@@ -1,34 +1,14 @@
 package com.eventstore.dbclient;
 
-import io.grpc.Metadata;
-
-public class UpdatePersistentSubscriptionOptions {
-    private final ConnectionMetadata metadata;
+public class UpdatePersistentSubscriptionOptions extends OptionsBase<UpdatePersistentSubscriptionOptions> {
     private PersistentSubscriptionSettings settings;
 
     private UpdatePersistentSubscriptionOptions() {
-        this.metadata = new ConnectionMetadata();
         this.settings = PersistentSubscriptionSettings.builder().build();
     }
 
     public static UpdatePersistentSubscriptionOptions get() {
         return new UpdatePersistentSubscriptionOptions();
-    }
-
-    public Metadata getMetadata() {
-        return this.metadata.build();
-    }
-
-    public UpdatePersistentSubscriptionOptions authenticated(UserCredentials credentials) {
-        if(credentials == null)
-            return this;
-
-        this.metadata.authenticated(credentials);
-        return this;
-    }
-
-    public boolean hasUserCredentials() {
-        return this.metadata.hasUserCredentials();
     }
 
     public PersistentSubscriptionSettings getSettings() {
